+++ CHANGELOG.md
@@ -3,6 +3,13 @@
 From 2019 onwards, all notable changes to tarpaulin will be documented in this
 file.
 
+## [0.35.3] 2026-04-25
+### Changed
+- Refined the forced coverage of generic items and expressions to reduce false negatives
+
+### Fixed
+- Fix ptrace SIGILL issue where PID parent event is delivered in iteration after the child stop is reported
+
 ## [0.35.2] 2026-02-19
 ### Fixed
 - Update native-tls to fix install issue after breaking change.